# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Lead Verification Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCurrentUser
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username or email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change current user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized or incorrect current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"

    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/start:
    post:
      operationId: startCampaign
      tags: [campaigns]
      summary: Start a campaign (begin calling)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /campaigns/{id}/pause:
    post:
      operationId: pauseCampaign
      tags: [campaigns]
      summary: Pause a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, name, email, role, companyId]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        companyId:
          type: integer
        companyName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        email:
          type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
...
    Campaign:
      type: object
      required: [id, name, status, timezone, startTime, endTime, callsPerMinute, concurrentCallLimit, retryAttempts, retryDelay, retryStrategy, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed]
        timezone:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        callsPerMinute:
          type: integer
        concurrentCallLimit:
          type: integer
        retryAttempts:
          type: integer
        retryDelay:
          type: integer
        retryStrategy:
          $ref: "#/components/schemas/RetryStrategy"
        totalLeads:
          type: integer
        processedLeads:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    CampaignInput:
      type: object
      required: [name, timezone, startTime, endTime, callsPerMinute, concurrentCallLimit, retryStrategy]
      properties:
        name:
          type: string
        timezone:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        callsPerMinute:
          type: integer
        concurrentCallLimit:
          type: integer
        retryAttempts:
          type: integer
        retryDelay:
          type: integer
        retryStrategy:
          $ref: "#/components/schemas/RetryStrategy"
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        timezone:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        callsPerMinute:
          type: integer
        concurrentCallLimit:
          type: integer
        retryAttempts:
          type: integer
        retryDelay:
          type: integer
        retryStrategy:
          $ref: "#/components/schemas/RetryStrategy"
    RetryRule:
      type: object
      required: [id, delayValue, delayUnit, timeRule]
      properties:
        id:
          type: string
        delayValue:
          type: integer
        delayUnit:
          type: string
          enum: [minutes, hours, days, weeks]
        timeRule:
          type: string
          enum: [same_day, next_day, same_time, different_time, business_hours_only]
        fixedTime:
          type: string
        timeOffsetHours:
          type: integer
    RetryStrategy:
      type: object
      required: [retryType, maxAttempts, businessDaysOnly, timeRotation, schedule, frequency, timePreference, allowedDays, stopConditions]
      properties:
        retryType:
          type: string
          enum: [fixed_delay, custom_schedule, smart_schedule]
        maxAttempts:
          type: integer
        businessDaysOnly:
          type: boolean
        timeRotation:
          type: boolean
        schedule:
          type: array
          items:
            $ref: "#/components/schemas/RetryRule"
        frequency:
          type: string
          enum: [same_day, next_day, alternate_day, every_x_days, weekly, monthly, custom_cron]
        frequencyDays:
          type: integer
        cronExpression:
          type: string
        timePreference:
          type: string
          enum: [same_time, random_in_window, fixed_time, different_time, ai_predicted]
        fixedRetryTime:
          type: string
        allowedDays:
          type: array
          items:
            type: integer
        stopConditions:
          type: array
          items:
            type: string
            enum: [answered, callback_requested, voicemail_detected, dnc, max_attempts, lead_status_changed]
...